+++ R/cmd-context.R
@@ -22,6 +22,7 @@ cmd_wd <- function(command, wd = NULL) {
 #'  - `cmd_envvar`: Named character define the environment variables.
 #'  - `cmd_envpath`: Unnamed character to define the `PATH`-like environment
 #' variables `name`.
+#'  - `cmd_on_start`: Expression to be evaluated when the command started.
 #'  - `cmd_on_exit`: Expression to be evaluated when the command finished.
 #' @param action Should the new values `"replace"`, `"prefix"` or `"suffix"`
 #' existing environment variables?
@@ -88,6 +89,16 @@ cmd_envpath <- function(command, ..., action = "prefix", name = "PATH") {
         action = action,
         sep = .Platform$path.sep
     )
+}
+
+#' @describeIn cmd_wd define the start code of the command
+#' @return
+#' - `cmd_on_start`: The `command` object itself, with the start code updated.
+#' @export
+cmd_on_start <- function(command, ...) {
+    assert_s3_class(command, "command")
+    command$on_start <- c(.subset2(command, "on_start"), rlang::enquos(...))
+    command
 }
 
 #' @describeIn cmd_wd define the exit code of the command